Transformers defined in item classes
- Moved transformer defintion into item type classes.
- Moved single methods into item type models as well as adjusting the queries.
- Updated delete to use
